+++ pwndbg/libc/__init__.py
@@ -1,8 +1,35 @@
 from __future__ import annotations
 
-from .api import Libc
-from .api import LibcType
-from .api import LibcURLs
-from .get import get
+from .dispatch import LibcType
+from .dispatch import LibcURLs
+from .dispatch import LibcWrangler
+from .facade import addr
+from .facade import filepath
+from .facade import has_debug_info
+from .facade import has_symbols
+from .facade import loader_addr
+from .facade import loader_filepath
+from .facade import relocations_by_section_name
+from .facade import section_address_by_name
+from .facade import section_by_name
+from .facade import urls
+from .facade import version
+from .facade import which
 
-__all__ = ["Libc", "LibcType", "LibcURLs", "get"]
+__all__ = [
+    "LibcWrangler",
+    "LibcType",
+    "LibcURLs",
+    "which",
+    "addr",
+    "filepath",
+    "has_debug_info",
+    "has_symbols",
+    "loader_addr",
+    "loader_filepath",
+    "relocations_by_section_name",
+    "section_address_by_name",
+    "section_by_name",
+    "urls",
+    "version",
+]
