+++ pwndbg/libc/musl.py
@@ -4,9 +4,8 @@ from pathlib import Path
 
 from elftools.elf.relocation import Relocation
 
-from . import common
-from .api import LibcType
-from .api import LibcURLs
+from .dispatch import LibcType
+from .dispatch import LibcURLs
 
 
 def type() -> LibcType:
@@ -36,7 +35,7 @@ def _is_being_used() -> bool:
 
 
 def version() -> tuple[int, ...]:
-    return (1, 1)
+    raise NotImplementedError
 
 
 def has_symbols() -> bool:
@@ -46,33 +45,12 @@ def has_symbols() -> bool:
 def has_debug_info() -> bool:
     return True
 
-
-def filepath() -> Path:
-    return Path("")
-
-
-def loader_filepath() -> Path:
-    return Path("")
+def verify_libc_candidate(mapping_name: str) -> bool:
+    ...
 
 
-def addr() -> int:
-    return 0
-
-
-def loader_addr() -> int:
-    return 0
-
-
-def section_by_name(section_name: str) -> tuple[int, int, bytes] | None:
-    return common.section_by_name(section_name, filepath())
-
-
-def section_address_by_name(section_name: str) -> int:
-    return common.section_address_by_name(section_name, filepath())
-
-
-def relocations_by_section_name(section_name: str) -> tuple[Relocation, ...]:
-    return common.relocations_by_section_name(section_name, filepath())
+def verify_ld_candidate(mapping_name: str) -> bool:
+    return verify_libc_candidate(mapping_name)
 
 
 def urls() -> LibcURLs:
